+++ server/proliferate/server/cloud/runtime/config_sync/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime configuration sync owners."""
