+++ server/proliferate/server/cloud/runtime/service.py
@@ -10,20 +10,20 @@ from proliferate.db.store.cloud_runtime_environments import load_runtime_environ
 from proliferate.db.store.cloud_workspaces import load_cloud_workspace_by_id
 from proliferate.server.billing.service import get_billing_snapshot_for_subject
 from proliferate.server.cloud.errors import CloudApiError
-from proliferate.server.cloud.runtime.anyharness_api import (
-    get_runtime_ready_agent_kinds,
+from proliferate.server.cloud.runtime.config_sync.worktree_policy import (
+    sync_cloud_worktree_policy_to_runtime,
 )
-from proliferate.server.cloud.runtime.auth_status import (
+from proliferate.server.cloud.runtime.credentials.auth_status import (
     load_workspace_runtime_auth_snapshot,
 )
-from proliferate.server.cloud.runtime.ensure_running import (
+from proliferate.server.cloud.runtime.credentials.remote_agents import (
+    get_runtime_ready_agent_kinds,
+)
+from proliferate.server.cloud.runtime.liveness.ensure_running import (
     ensure_environment_runtime_ready,
 )
 from proliferate.server.cloud.runtime.models import RuntimeConnectionTarget
 from proliferate.server.cloud.runtime.provision import provision_workspace as _provision_workspace
-from proliferate.server.cloud.runtime.worktree_policy_sync import (
-    sync_cloud_worktree_policy_to_runtime,
-)
 from proliferate.utils.crypto import decrypt_text
 
 provision_workspace = _provision_workspace
